requirements.txt
bleach==3.1.0
certifi==2019.11.28
chardet==3.0.4
cloudinary==1.19.1
Django==2.2.10
django-cleanup==4.0.0
django-cloudinary-storage==0.2.3
django-js-asset==1.2.2
django-mptt==0.11.0
ebaysdk==2.1.5
geoip2==3.0.0
gunicorn==20.0.4
idna==2.8
lxml==4.5.0
maxminddb==1.5.2
mock==4.0.1
Pillow==7.0.0
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.0.1
